requirements.txt
pymongo[srv]
requests
fastapi 
pymongo 
uvicorn
python-decouple
python-multipart

